+++ src/agentos/engine/steps/__init__.py
@@ -1,6 +1,7 @@
 """Pre-turn pipeline steps."""
 
 from agentos.engine.pipeline import TurnContext
+from agentos.engine.steps.inject_env_probe import inject_env_probe
 from agentos.engine.steps.inject_platform_hint import inject_platform_hint
 from agentos.engine.steps.inject_subagent_grounding import inject_subagent_grounding
 from agentos.engine.steps.prompt_cache import apply_prompt_cache
@@ -20,6 +21,7 @@ __all__ = [
     "apply_prompt_cache",
     "apply_agentos_router",
     "filter_skills",
+    "inject_env_probe",
     "inject_platform_hint",
     "inject_subagent_grounding",
     "observe_reasoning_hint",